Throw when reassigning baseUrl after normalization
