+++ test/create.js
@@ -164,87 +164,3 @@ test('defaults are cloned on instance creation', t => {
 	t.not(options.foo, instance.defaults.options.foo);
 	t.not(methods[0], instance.defaults.methods[0]);
 });
-
-test('merging instances', async t => {
-	const instanceA = got.extend({headers: {unicorn: 'rainbow'}});
-	const instanceB = got.extend({baseUrl: s.url});
-	const merged = got.mergeInstances(instanceA, instanceB, ['get']);
-
-	const headers = (await merged('/', {json: true})).body;
-	t.is(headers.unicorn, 'rainbow');
-	t.not(headers['user-agent'], undefined);
-});
-
-test('merging already merged instances & another instance', async t => {
-	const instanceA = got.extend({headers: {dog: 'woof'}});
-	const instanceB = got.extend({headers: {cat: 'meow'}});
-	const instanceC = got.extend({headers: {bird: 'tweet'}});
-	const instanceD = got.extend({headers: {mouse: 'squeek'}});
-
-	const merged = got.mergeInstances(instanceA, instanceB, instanceC);
-	const doubleMerged = got.mergeInstances(merged, instanceD);
-
-	const headers = (await doubleMerged(s.url, {json: true})).body;
-	t.is(headers.dog, 'woof');
-	t.is(headers.cat, 'meow');
-	t.is(headers.bird, 'tweet');
-	t.is(headers.mouse, 'squeek');
-});
-
-test('merging two groups of merged instances', async t => {
-	const instanceA = got.extend({headers: {dog: 'woof'}});
-	const instanceB = got.extend({headers: {cat: 'meow'}});
-	const instanceC = got.extend({headers: {bird: 'tweet'}});
-	const instanceD = got.extend({headers: {mouse: 'squeek'}});
-
-	const groupA = got.mergeInstances(instanceA, instanceB);
-	const groupB = got.mergeInstances(instanceC, instanceD);
-
-	const merged = got.mergeInstances(groupA, groupB);
-
-	const headers = (await merged(s.url, {json: true})).body;
-	t.is(headers.dog, 'woof');
-	t.is(headers.cat, 'meow');
-	t.is(headers.bird, 'tweet');
-	t.is(headers.mouse, 'squeek');
-});
-
-test('hooks are merged when merging instances', t => {
-	const getBeforeRequestHooks = instance => instance.defaults.options.hooks.beforeRequest;
-
-	const instanceA = got.extend({hooks: {
-		beforeRequest: [
-			options => {
-				options.headers.dog = 'woof';
-			}
-		]
-	}});
-	const instanceB = got.extend({hooks: {
-		beforeRequest: [
-			options => {
-				options.headers.cat = 'meow';
-			}
-		]
-	}});
-
-	const merged = got.mergeInstances(instanceA, instanceB);
-	t.deepEqual(getBeforeRequestHooks(merged), getBeforeRequestHooks(instanceA).concat(getBeforeRequestHooks(instanceB)));
-});
-
-test('hooks can be passed by when merging instances', t => {
-	const instanceA = got.extend({hooks: {
-		beforeRequest: [
-			options => {
-				options.headers.dog = 'woof';
-			}
-		]
-	}});
-	const instanceB = got.create({
-		methods: [],
-		options: {
-			hooks: {}
-		}
-	});
-	const merged = got.mergeInstances(instanceA, instanceB);
-	t.deepEqual(merged.defaults.options.hooks.beforeRequest, instanceA.defaults.options.hooks.beforeRequest);
-});
